requirements.txt
# Install transformers from github
git+https://github.com/huggingface/transformers.git
# transformers
# Install torch version in that matches your env
--extra-index-url https://download.pytorch.org/whl/cu113
torch
# Check README for details in installing deepspeed
# deepspeed
hydra-core
datasets
tqdm
scipy
sklearn
tensorboard